renamed serversStatus to serverStatus
