+++ src/api/index.ts
@@ -2,7 +2,7 @@ import express from 'express'
 import adminRouter from './admin'
 import mappingRouter from './mapping'
 import sshKeyRouter from './sshKeys'
-import statusRouter from './serversStatus'
+import statusRouter from './serverStatus'
 import { getAvailableDomains } from '../lib/data'
 
 const apiRouter = express.Router()
